+++ src/surrealist/ql/statements/define_index_statements.py
@@ -1,6 +1,8 @@
 from surrealist.ql.statements.statement import FinishedStatement, Statement
 
 
+# This module contains classes with duplicate code. This is made specifically for IDE hints
+
 class Unique(FinishedStatement):
     """
     Represents UNIQUE statement
@@ -34,7 +36,8 @@ class SearchAnalyzer(FinishedStatement):
 
 class MTree(FinishedStatement):
     """
-    Represents MTREE index
+    Represents MTREE index.
+    There is code duplication in this class. This is made specifically for IDE hints
 
     Refer to:
     https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#vector-search-indexes
@@ -47,28 +50,133 @@ class MTree(FinishedStatement):
         self._capacity = None
         self._type = None
 
-    def type(self, mtree_type: str) -> "MTree":
-        self._type = mtree_type
-        return self
-
     def distance_euclidean(self) -> "MTree":
+        """
+        Make index to use Euclidean distance.
+        When no function is specified, SurrealDB chooses Euclidean for the default distance function
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/functions/vector#vectordistanceeuclidean
+
+        :return: MTree object
+        """
         self._dist = "EUCLIDEAN"
         return self
 
     def distance_manhattan(self) -> "MTree":
+        """
+        Make index to use Manhattan distance
+        When no function is specified, SurrealDB chooses Euclidean for the default distance function
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/functions/vector#vectordistancemanhattan
+
+        :return: MTree object
+        """
         self._dist = "MANHATTAN"
         return self
 
     def distance_cosine(self) -> "MTree":
+        """
+        Make index to use Cosine distance
+        When no function is specified, SurrealDB chooses Euclidean for the default distance function
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/functions/vector#vectorsimilaritycosine
+
+        :return: MTree object
+        """
         self._dist = "COSINE"
         return self
 
     def distance_minkowski(self) -> "MTree":
+        """
+        Make index to use Minkowski distance
+        When no function is specified, SurrealDB chooses Euclidean for the default distance function
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/functions/vector#vectordistanceminkowski
+
+        :return: MTree object
+        """
         self._dist = "MINKOWSKI"
         return self
 
-    def capacity(self, capacity_value: str) -> "MTree":
-        self._capacity = capacity_value
+    def capacity(self, value: int) -> "MTree":
+        """
+        The CAPACITY clause is used to specify the maximum number of records that can be stored in the index.
+        This is useful when you want to limit the number of records stored in the index to optimize performance
+        The default value is 40 (on SurrealDB side)
+
+        :param value: integer value for capacity
+        :return: MTree object
+        """
+        self._capacity = value
+        return self
+
+    def type_f64(self) -> "MTree":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 64-bit floating-point numbers (double precision floating-point numbers).
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: MTree object
+        """
+        self._type = "F64"
+        return self
+
+    def type_f32(self) -> "MTree":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 32-bit floating-point numbers (single precision floating-point numbers).
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: MTree object
+        """
+        self._type = "F32"
+        return self
+
+    def type_i64(self) -> "MTree":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 64-bit signed integers.
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: MTree object
+        """
+        self._type = "I64"
+        return self
+
+    def type_i32(self) -> "MTree":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 32-bit signed integers.
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: MTree object
+        """
+        self._type = "I32"
+        return self
+
+    def type_i16(self) -> "MTree":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 16-bit signed integers.
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: MTree object
+        """
+        self._type = "I16"
         return self
 
     def _clean_str(self):
@@ -81,6 +189,7 @@ class MTree(FinishedStatement):
 class HNSW(FinishedStatement):
     """
     Represents HNSW index
+    There is code duplication in this class. This is made specifically for IDE hints
 
     Refer to:
     https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#hnsw-hierarchical-navigable-small-world-since-150
@@ -94,31 +203,147 @@ class HNSW(FinishedStatement):
         self._type = None
         self._m = None
 
-    def type(self, mtree_type: str) -> "HNSW":
-        self._type = mtree_type
+    def type_f64(self) -> "HNSW":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 64-bit floating-point numbers (double precision floating-point numbers).
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: HNSW object
+        """
+        self._type = "F64"
+        return self
+
+    def type_f32(self) -> "HNSW":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 32-bit floating-point numbers (single precision floating-point numbers).
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: HNSW object
+        """
+        self._type = "F32"
+        return self
+
+    def type_i64(self) -> "HNSW":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 64-bit signed integers.
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: HNSW object
+        """
+        self._type = "I64"
+        return self
+
+    def type_i32(self) -> "HNSW":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 32-bit signed integers.
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: HNSW object
+        """
+        self._type = "I32"
+        return self
+
+    def type_i16(self) -> "HNSW":
+        """
+        The TYPE clause is optional and can be used to specify the data type of the vector.
+        Represents 16-bit signed integers.
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#types
+
+        :return: HNSW object
+        """
+        self._type = "I16"
         return self
 
     def distance_euclidean(self) -> "HNSW":
+        """
+        Make index to use Euclidean distance
+        When no function is specified, SurrealDB chooses Euclidean for the default distance function
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/functions/vector#vectordistanceeuclidean
+
+        :return: HNSW object
+        """
         self._dist = "EUCLIDEAN"
         return self
 
     def distance_manhattan(self) -> "HNSW":
+        """
+        Make index to use Manhattan distance
+        When no function is specified, SurrealDB chooses Euclidean for the default distance function
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/functions/vector#vectordistancemanhattan
+
+        :return: HNSW object
+        """
         self._dist = "MANHATTAN"
         return self
 
     def distance_cosine(self) -> "HNSW":
+        """
+        Make index to use Cosine distance
+        When no function is specified, SurrealDB chooses Euclidean for the default distance function
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/functions/vector#vectorsimilaritycosine
+
+        :return: HNSW object
+        """
         self._dist = "COSINE"
         return self
 
     def distance_minkowski(self) -> "HNSW":
+        """
+        Make index to use Minkowski distance
+        When no function is specified, SurrealDB chooses Euclidean for the default distance function
+
+        Refer to:
+        https://surrealdb.com/docs/surrealdb/surrealql/functions/vector#vectordistanceminkowski
+
+        :return: HNSW object
+        """
         self._dist = "MINKOWSKI"
         return self
 
-    def efc(self, value: str) -> "HNSW":
+    def efc(self, value: int) -> "HNSW":
+        """
+        Exploration factor during construction. This parameter determines the size of the dynamic list for the nearest
+        neighbor candidates during the graph construction phase. A larger efConstruction value leads to a more thorough
+        construction, improving the quality and accuracy of the search but increasing construction time.
+        The default value is 150 (on SurrealDB side)
+
+        :param value: integer value for EFC
+        :return: HNSW object
+        """
         self._efc = value
         return self
 
-    def m(self, value: str) -> "HNSW":
+    def max_connections(self, value: int) -> "HNSW":
+        """
+        Max Connections per Element. Defines the maximum number of bidirectional links (neighbors) per node in each
+        layer of the graph, except for the lowest layer. This parameter controls the connectivity and overall structure
+        of the network. Higher values of MM generally improve search accuracy but increase memory usage and
+        construction time.
+        The default value is 12 (on the SurrealDB side)
+
+        :param value: integer value for MM
+        :return: HNSW object
+        """
         self._m = value
         return self
 
@@ -158,7 +383,7 @@ class CanUseIndexTypes:
 
     def mtree(self, num_dimension: int) -> MTree:
         """
-        Creates vector-search index
+        Creates vector-search index MTREE
 
         Refer to:
         https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#vector-search-indexes
@@ -170,12 +395,12 @@ class CanUseIndexTypes:
 
     def hnsw(self, num_dimension: int) -> HNSW:
         """
-        Creates vector-search index
+        Creates vector-search index HNSW
 
         Refer to:
         https://surrealdb.com/docs/surrealdb/surrealql/statements/define/indexes#hnsw-hierarchical-navigable-small-world-since-150
 
         :param num_dimension: number of dimensions
-        :return: MTree object
+        :return: HNSW object
         """
         return HNSW(self, num_dimension)