+++ tests/unit_tests/test_database.py
@@ -101,6 +101,8 @@ SIGNIN (SELECT * FROM user WHERE email = $email AND crypto::argon2::compare(pass
         self.assertEqual(text, DefineIndex(None, "userNameIndex", "user").columns("name").mtree(4).distance_minkowski().to_str())
         text = "DEFINE INDEX userNameIndex ON TABLE user COLUMNS name MTREE DIMENSION 4 DIST COSINE;"
         self.assertEqual(text, DefineIndex(None, "userNameIndex", "user").columns("name").mtree(4).distance_cosine().to_str())
+        text = "DEFINE INDEX userNameIndex ON TABLE user COLUMNS name MTREE DIMENSION 3 TYPE F64 CAPACITY 50;"
+        self.assertEqual(text, DefineIndex(None, "userNameIndex", "user").columns("name").mtree(3).type_f64().capacity(50).to_str())
 
     def test_define_index_hnsw(self):
         text = "DEFINE INDEX userNameIndex ON TABLE user COLUMNS name HNSW DIMENSION 4;"
@@ -114,7 +116,7 @@ SIGNIN (SELECT * FROM user WHERE email = $email AND crypto::argon2::compare(pass
         text = "DEFINE INDEX userNameIndex ON TABLE user COLUMNS name HNSW DIMENSION 4 DIST COSINE;"
         self.assertEqual(text, DefineIndex(None, "userNameIndex", "user").columns("name").hnsw(4).distance_cosine().to_str())
         text = "DEFINE INDEX userNameIndex ON TABLE user COLUMNS name HNSW DIMENSION 4 DIST COSINE EFC 150 M 2;"
-        self.assertEqual(text, DefineIndex(None, "userNameIndex", "user").columns("name").hnsw(4).distance_cosine().efc("150").m(2).to_str())
+        self.assertEqual(text, DefineIndex(None, "userNameIndex", "user").columns("name").hnsw(4).distance_cosine().efc(150).max_connections(2).to_str())
 
     def test_define_index_exists(self):
         text = "DEFINE INDEX IF NOT EXISTS userNameIndex ON TABLE user COLUMNS name SEARCH ANALYZER ascii " \